Sanity check for TSO payload length in TCP chopper.
Although it should not happen, check that ph_mss is not 0 in
tcp_chopper().  This could catch errors in the LRO path of network
drivers.  Better count bad packet and drop it rather than ending
in an endless loop.  The new logic is analog to a recent change in
the hardware TSO path in the drivers.

OK jan@