repair qwx(4) WEP and TKIP via software crypto
It is difficult to make WEP and WPA1/TKIP work with hardware crypto.
Add a comment which explains why.

Ensure that setkey task state is properly cleared when the interface
goes down. This issue was found while trying to add WEP keys for hw
crypto, but is still worth fixing in general.

Also, use m_makespace to append trailing padding for the MIC when
hardware crypto is used in combination with "raw" frame mode (not
the default), instead of blindly adjusting m_len.

ok kevlo@